Class_Assignments/01-Multi_Models_Showdown/pyproject.toml
[project]
name = "01-multi-models-showdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.0.16",
    "streamlit>=1.45.1",
]
